refactor: moved form elements to UI components
